Separate input stream unit tests into two
